requirements.txt
azure-ai-evaluation
azure-identity
azure-ai-projects
openai
python-dotenv
promptflow
promptflow-azure
promptflow-core
notebook
ipykernel
pandas
typing-extensions